DL/requirements.txt
streamlit>=1.32.0
mediapipe>=0.10.9
opencv-python-headless>=4.9.0
numpy>=1.24.0
Pillow>=10.0.0
pandas>=2.0.0
pillow-heif>=0.15.0
